requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.1
flask==3.0.3
flask-sqlalchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
idna==3.7
importlib-metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
psycopg2-binary==2.9.10
pydantic==2.7.1
pydantic-core==2.18.2
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.37.2
typing-extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
werkzeug==3.0.2
zipp==3.18.1
